+++ lib/address_autocomplete_widgets.dart
@@ -1,0 +1,6 @@
+library google_maps_places_autocomplete_widgets;
+
+export 'package:google_maps_places_autocomplete_widgets/widgets/address_autocomplete_textfield.dart';
+export 'package:google_maps_places_autocomplete_widgets/widgets/address_autocomplete_textformfield.dart';
+export 'package:google_maps_places_autocomplete_widgets/model/place.dart';
+export 'package:google_maps_places_autocomplete_widgets/model/suggestion.dart';