+++ test/google_maps_places_autocomplete_widgets_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:maps_places_autocomplete/maps_places_autocomplete.dart';
+import 'package:google_maps_places_autocomplete_widgets/address_autocomplete_widgets.dart';
 
 void main() {
   // test('adds one to input values', () {